+++ planck-c/functions.h
@@ -135,3 +135,6 @@ JSValueRef function_socket_write(JSContextRef ctx, JSObjectRef function, JSObjec
 
 JSValueRef function_socket_close(JSContextRef ctx, JSObjectRef function, JSObjectRef thisObject,
                                  size_t argc, const JSValueRef args[], JSValueRef *exception);
+
+JSValueRef function_sleep(JSContextRef ctx, JSObjectRef function, JSObjectRef thisObject,
+                          size_t argc, const JSValueRef args[], JSValueRef *exception);